chore: upgrade hyper nano image
