+++ .github/workflows/publish-image-nano.yml
@@ -1,18 +1,20 @@
 name: publish nano
+
 on:
   push:
     branches:
       - "main"
     paths:
       - "images/nano/**"
+
 jobs:
   build:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        deno-version: [1.13.x]
+        deno-version: [1.x]
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Use Deno ${{ matrix.deno-version }}
         uses: denoland/setup-deno@v1
         with:
