+++ images/nano/.gitignore
@@ -1,2 +1,3 @@
 data
 .env
+hyper
