+++ images/nano/deps.js
@@ -1,10 +1,10 @@
-export { default as hyper } from "https://x.nest.land/hyper@2.0.0/mod.js";
-export { default as app } from "https://x.nest.land/hyper-app-opine@1.2.10/mod.js";
+export { default as hyper } from "https://x.nest.land/hyper@3.1.0/mod.js";
+export { default as app } from "https://x.nest.land/hyper-app-opine@2.0.1/mod.js";
 
 // adapters
 export { default as dndb } from "https://x.nest.land/hyper-adapter-dndb@2.0.0/mod.js";
-export { default as cache } from "https://x.nest.land/hyper-adapter-sqlite@1.0.0/mod.js";
-export { default as fs } from "https://x.nest.land/hyper-adapter-fs@1.0.8/mod.js";
-export { default as minisearch } from "https://x.nest.land/hyper-adapter-minisearch@1.1.1/mod.js";
+export { default as cache } from "https://x.nest.land/hyper-adapter-sqlite@2.0.0/mod.js";
+export { default as fs } from "https://x.nest.land/hyper-adapter-fs@2.0.2/mod.js";
+export { default as minisearch } from "https://x.nest.land/hyper-adapter-minisearch@2.0.0/mod.js";
 // include queue locally
-export { default as queue } from "https://x.nest.land/hyper-adapter-queue@0.0.3/mod.js";
+export { default as queue } from "https://x.nest.land/hyper-adapter-queue@0.0.4/mod.js";
