+++ images/nano/main.js
@@ -1,0 +1,9 @@
+import { hyper } from "./deps.js";
+import config from "./hyper.config.js";
+
+export const main = (middleware = []) => {
+  return hyper({
+    ...config,
+    middleware,
+  });
+};
