--- images/nano/mod_test.js
@@ -1,23 +1,0 @@
-// import {
-//   assert,
-//   assertEquals,
-//   equal,
-// } from "https://deno.land/std@0.106.0/testing/asserts.ts";
-import { superdeno } from "https://deno.land/x/superdeno@4.4.0/mod.ts";
-import { hyper } from "./deps.js";
-import config from "./hyper.config.js";
-
-const test = Deno.test;
-
-test({
-  name: "GET /",
-  async fn() {
-    const app = await hyper(config);
-    await superdeno(app)
-      .get("/")
-      .expect("Content-Type", /json/)
-      .expect(200);
-  },
-  sanitizeResources: false,
-  sanitizeOps: false,
-});
