+++ images/nano/scripts/test.sh
@@ -1,1 +1,1 @@
-deno lint && deno fmt --check && DENO_ENV=test deno test -A --unstable --no-check=remote mod_test.js
+deno lint && deno fmt --check && DENO_ENV=test deno test -A --unstable --no-check=remote
